Permissions Menu

+++ routes/master/bank.php
@@ -4,9 +4,9 @@ use App\Http\Controllers\Master\BankController;
 use Illuminate\Support\Facades\Route;
 
 Route::prefix("bank")->group(function() {
-    Route::get("/", [BankController::class, "index"]);
-    Route::post("/", [BankController::class, "store"]);
-    Route::get("/{id}/edit", [BankController::class, "edit"]);
-    Route::put("/{id}", [BankController::class, "update"]);
-    Route::delete("/{id}", [BankController::class, "destroy"]);
+    Route::get("/", [BankController::class, "index"])->middleware("permission:bank.read");
+    Route::post("/", [BankController::class, "store"])->middleware("permission:bank.create");
+    Route::get("/{id}/edit", [BankController::class, "edit"])->middleware("permission:bank.edit");
+    Route::put("/{id}", [BankController::class, "update"])->middleware("permission:bank.edit");
+    Route::delete("/{id}", [BankController::class, "destroy"])->middleware("permission:bank.delete");
 });
